requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
astroid==2.11.5
asyncpg==0.25.0
autopep8==1.6.0
boto3==1.24.4
botocore==1.27.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.2
databases==0.5.5
dill==0.3.5.1
fastapi==0.78.0
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
isort==5.10.1
jmespath==1.0.0
lazy-object-proxy==1.7.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
passlib==1.7.4
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
types-requests==2.27.30
types-urllib3==1.26.15
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
wrapt==1.14.1
